+++ LabManagementBackend/LabManagement.DAL/Models/Booking.cs
@@ -17,7 +17,7 @@ public partial class Booking
 
     public DateTime EndTime { get; set; }
 
-    public decimal Status { get; set; }
+    public int Status { get; set; }
 
     public DateTime CreatedAt { get; set; }
